+++ ads/aqua/modeldeployment/constants.py
@@ -12,4 +12,3 @@ This module contains constants used in Aqua Model Deployment.
 DEFAULT_WAIT_TIME = 12000
 DEFAULT_POLL_INTERVAL = 10
 
-SHAPE_MAP = {"NVIDIA_GPU": "GPU"}
